fix: move sanitizeErrorMessage to validation module to fix CI timeout
The function was defined in index.ts which starts the MCP server on
import. Tests importing it caused the server to start on stdin,
hanging the process indefinitely. Moving it to validation.ts (a
side-effect-free module) fixes the 2-minute CI timeout.
